refactor: cast channelId to `uint256` to concatenate idx correctly

+++ contracts/LSP6KeyManager/LSP6KeyManagerCore.sol
@@ -74,7 +74,7 @@ abstract contract LSP6KeyManagerCore is ERC165, ILSP6KeyManager {
      */
     function getNonce(address from, uint128 channelId) public view returns (uint256) {
         uint256 nonceInChannel = _nonceStore[from][channelId];
-        return (channelId << 128) | nonceInChannel;
+        return (uint256(channelId) << 128) | nonceInChannel;
     }
 
     /**
